Moved Celeba model
