+++ accdfl/core/datasets/Celeba.py
@@ -5,9 +5,7 @@ from collections import defaultdict
 
 import numpy as np
 import torch
-import torch.nn.functional as F
 from PIL import Image
-from torch import nn
 from torch.nn import CrossEntropyLoss
 from torch.utils.data import DataLoader
 
@@ -16,7 +14,6 @@ from accdfl.core.datasets.Data import Data
 from accdfl.core.datasets.Dataset import Dataset
 from accdfl.core.datasets.Partitioner import DataPartitioner
 from accdfl.core.mappings.Mapping import Mapping
-from accdfl.core.models.Model import Model
 
 IMAGE_DIM = 84
 CHANNELS = 3
@@ -410,48 +407,3 @@ class Celeba(Dataset):
         loss_val = loss_val / count
         logging.info("Overall accuracy is: {:.1f} %".format(accuracy))
         return accuracy, loss_val
-
-
-class CNN(Model):
-    """
-    Class for a CNN Model for Celeba
-
-    """
-
-    def __init__(self):
-        """
-        Constructor. Instantiates the CNN Model
-            with 84*84*3 Input and 2 output classes
-
-        """
-        super().__init__()
-        # 2.8k parameters
-        self.conv1 = nn.Conv2d(CHANNELS, 32, 3, padding="same")
-        self.pool = nn.MaxPool2d(2, 2)
-        self.conv2 = nn.Conv2d(32, 32, 3, padding="same")
-        self.conv3 = nn.Conv2d(32, 32, 3, padding="same")
-        self.conv4 = nn.Conv2d(32, 32, 3, padding="same")
-        self.fc1 = nn.Linear(5 * 5 * 32, NUM_CLASSES)
-
-    def forward(self, x):
-        """
-        Forward pass of the model
-
-        Parameters
-        ----------
-        x : torch.tensor
-            The input torch tensor
-
-        Returns
-        -------
-        torch.tensor
-            The output torch tensor
-
-        """
-        x = F.relu(self.pool(self.conv1(x)))
-        x = F.relu(self.pool(self.conv2(x)))
-        x = F.relu(self.pool(self.conv3(x)))
-        x = F.relu(self.pool(self.conv4(x)))
-        x = torch.flatten(x, 1)
-        x = self.fc1(x)
-        return x
